feat(messages): implement emoji reactions UI with quick picker
- Add reaction pills below message bubbles showing emoji + count
- Implement click-to-toggle on reaction pills (add if not reacted, remove if already reacted)
- Add "React..." menu item to message context menu
- Create quick reaction picker with 8 common emojis (❤️🔥😂😮😢🙏👍👏)
- Integrate with getstream.io reactions API (addReaction/removeReaction)
- Cache reaction pill bounds for hit testing
- Highlight pills where current user has reacted with lighter background

Also fix compilation errors:
- Fix JUCE FontOptions::withStyle() to use String instead of FontStyleFlags
  (updated EditProfile.cpp and Profile.cpp to use "Bold" instead of juce::Font::bold)
- Fix [[maybe_unused]] attribute placement in API client files
  (moved before 'static' keyword for correct C++ syntax)

Task 6.6.4 from messaging enhancement plan.

+++ plugin/src/network/api/DownloadClient.cpp
@@ -111,7 +111,7 @@ void NetworkClient::downloadFile(const juce::String& url, const juce::File& targ
                 int bytesReadThisChunk;
                 while ((bytesReadThisChunk = stream->read(buffer.getData(), bufferSize)) > 0)
                 {
-                    output.write(buffer.getData(), bytesReadThisChunk);
+                    output.write(buffer.getData(), static_cast<size_t>(bytesReadThisChunk));
                     bytesRead += bytesReadThisChunk;
 
                     // Throttled progress reporting
